moves classes to their own file
